Update ListItem.js
Properly uses states. Also fixed the weird issue where the state was backwards from what happened (basically before, it did the calculations based on what the state was before it was clicked. Now it does it afterwards, which is the proper way and is a lot less confusing to follow).