Update App.css
Everything is now properly laid out in their own appropriate containers. Tables will now no longer overflow over the edge of the screen and will instead allow you to scroll to see the whole thing. Also played around with the <body> styling a bit to remove the ugly white border and shortened some of the margins a bit.